requirements.txt
google-cloud-firestore
google-cloud-language
feedparser
bleach
